requirements.txt
Sphinx
pyyaml
cryptography